fix add account bug

+++ plugins/account_checker/index.js
@@ -60,7 +60,7 @@ const isExpired = (server, account) => {
     account.expireTime = expireTime;
     if(expireTime <= Date.now() || data.create >= Date.now()) {
       const nextCheckTime = 10 * 60 * 1000 + randomInt(30000);
-      if(account.autoRemove && expireTime + account.autoRemoveDelay < Date.now()) {
+      if(account.active && account.autoRemove && expireTime + account.autoRemoveDelay < Date.now()) {
         modifyAccountFlow(server.id, account.id, nextCheckTime > account.autoRemoveDelay ? account.autoRemoveDelay : nextCheckTime);
         knex('account_plugin').delete().where({ id: account.id }).then();
       } else {
